// src/interfaces/moviesInterface.tsx
export interface IMovies {
  page: number;
  results: Result[];
  totalPages: number;
  totalResults: number;
}

export interface Result {
  adult: boolean;
  backdropPath: string;
  genreIds: number[];
  id: number;
  originalLanguage: string;
  originalTitle: string;
  overview: string;
  popularity: number;
  posterPath: string;
  releaseDate: string;
  title: string;
  video: boolean;
  voteAverage: number;
  voteCount: number;
}